Refactor out _prepare_for_flox
